Clean up unused, move PrimaryKey
